# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de E-commerce",
    "version": "1.0.0",
    "description": "API backend para gestión de productos y autenticación"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo",
      "security": [
        {
          "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }
      ]
    },
    {
      "url": "https://estudio-glow.onrender.com/",
      "description": "Servidor de producción",
      "security": [
        {
          "bearerAuth": {
            "type": "https",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }
      ]
    }
  ],
  "components": {
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "admin-user-001",
            "description": "ID del usuario"
          },
          "name": {
            "type": "string",
            "example": "Admin Estudio Glow",
            "description": "Nombre del usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "yasitacardenas3637@gmail.com",
            "description": "Correo electrónico del usuario"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "customer"
            ],
            "example": "admin",
            "description": "Rol del usuario"
          },
          "provider": {
            "type": "string",
            "enum": [
              "LOCAL",
              "GOOGLE"
            ],
            "example": "LOCAL",
            "description": "Proveedor de autenticación"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful",
            "description": "Mensaje de respuesta"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error message",
            "description": "Mensaje de error"
          }
        },
        "required": [
          "message"
        ]
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "ID del producto"
          },
          "name": {
            "type": "string",
            "example": "Laptop Gaming",
            "description": "Nombre del producto"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Laptop de alto rendimiento para gaming",
            "description": "Descripción del producto"
          },
          "price": {
            "type": "integer",
            "example": 1500,
            "description": "Precio del producto en ARS (entero, sin decimales)."
          },
          "stock": {
            "type": "number",
            "example": 50,
            "description": "Cantidad en stock"
          },
          "category": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "example": "550e8400-e29b-41d4-a716-446655440000",
                "description": "ID de la categoría"
              },
              "name": {
                "type": "string",
                "example": "Electrónica",
                "description": "Nombre de la categoría"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "example": {
              "id": "550e8400-e29b-41d4-a716-446655440000",
              "name": "Electrónica"
            },
            "description": "Categoría del producto (null si no tiene categoría asignada)"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/image.jpg",
            "description": "URL de la imagen del producto"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15T10:30:00Z",
            "description": "Fecha de creación"
          },
          "updatedAt": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15T10:30:00Z",
            "description": "Fecha de actualización"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "stock",
          "category",
          "imageUrl",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Laptop Gaming",
            "description": "Nombre del producto"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "example": "Laptop de alto rendimiento para gaming",
            "description": "Descripción del producto"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "example": 1500,
            "description": "Precio del producto en ARS (entero, sin decimales)."
          },
          "stock": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "example": 50,
            "description": "Cantidad en stock"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "ID de la categoría del producto"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Imagen del producto"
          }
        }
      },
...
      "CartItem": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "ID del producto"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "example": 1,
            "description": "Cantidad del producto"
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/products/{id}": {
      "get": {
        "tags": [
          "Products (Admin)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "1",
              "description": "ID del producto"
            },
            "required": true,
            "description": "ID del producto",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products (Admin)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "1",
              "description": "ID del producto"
            },
            "required": true,
            "description": "ID del producto",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error en la solicitud",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products (Admin)"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "1",
              "description": "ID del producto"
            },
            "required": true,
            "description": "ID del producto",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/checkout/mercadopago": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Iniciar pago con MercadoPago",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "description": "Clave de idempotencia requerida para evitar órdenes duplicadas (UUID o string único por intento de pago)"
            },
            "required": true,
            "description": "Clave de idempotencia requerida para evitar órdenes duplicadas (UUID o string único por intento de pago)",
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CartItem"
                    }
                  }
                },
                "required": [
                  "items"
                ],
                "example": {
                  "items": [
                    {
                      "productId": "550e8400-e29b-41d4-a716-446655440000",
                      "quantity": 2
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferencia de pago de MercadoPago creada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "preferenceUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "required": [
                    "orderId",
                    "preferenceUrl"
                  ],
                  "example": {
                    "orderId": "550e8400-e29b-41d4-a716-446655440000",
                    "preferenceUrl": "https://www.mercadopago.com.ar/checkout/v1/redirect?pref_id=..."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Carrito vacío, items inválidos o falta Idempotency-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Solicitud duplicada en vuelo (mismo Idempotency-Key concurrente)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "example": {
                    "error": {
                      "message": "Duplicate request in flight"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/checkout/transfer": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Iniciar pago por transferencia bancaria con comprobante",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "description": "Clave de idempotencia requerida para evitar órdenes duplicadas (UUID o string único por intento de pago)"
            },
            "required": true,
            "description": "Clave de idempotencia requerida para evitar órdenes duplicadas (UUID o string único por intento de pago)",
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "receipt": {
                    "nullable": true,
                    "description": "Imagen del comprobante de transferencia (JPG, PNG, WEBP — máx 5 MB)",
                    "example": "(binary file)"
                  },
                  "items": {
                    "type": "string",
                    "description": "JSON serializado del array de items del carrito",
                    "example": "[{\"productId\":\"550e8400-e29b-41d4-a716-446655440000\",\"quantity\":1}]"
                  }
                },
                "required": [
                  "items"
                ],
                "example": {
                  "receipt": "(binary file)",
                  "items": "[{\"productId\":\"550e8400-e29b-41d4-a716-446655440000\",\"quantity\":1}]"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orden de transferencia creada, pendiente de verificación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING_VERIFICATION"
                      ]
                    },
                    "receiptUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "required": [
                    "orderId",
                    "status",
                    "receiptUrl"
                  ],
                  "example": {
                    "orderId": "550e8400-e29b-41d4-a716-446655440000",
                    "status": "PENDING_VERIFICATION",
                    "receiptUrl": "https://res.cloudinary.com/example/receipts/abc123.jpg"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Comprobante faltante, carrito vacío, items inválidos o falta Idempotency-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Solicitud duplicada en vuelo (mismo Idempotency-Key concurrente)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "example": {
                    "error": {
                      "message": "Duplicate request in flight"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}